+++ gcovr/configuration.py
@@ -31,6 +31,8 @@ import re
 from . import formats
 from .options import (
     GcovrConfigOption,
+    GcovrConfigOptionAction,
+    GcovrDeprecatedConfigOptionAction,
     Options,
     OutputOrDefault,
     check_input_file,
@@ -166,7 +168,9 @@ def parse_config_into_dict(
             raise cfg_entry.error("unknown config option") from None
 
         value = _get_value_from_config_entry(cfg_entry, option)
-        _assign_value_to_dict(cfg_dict, value, option, is_single_value=True)
+        _assign_value_to_dict(
+            cfg_dict, value, option, cfg_entry_key=cfg_entry.key, is_single_value=True
+        )
 
     return cfg_dict
 
@@ -268,6 +272,7 @@ def _assign_value_to_dict(
     value: Any,
     option: GcovrConfigOption,
     is_single_value: bool,
+    cfg_entry_key: str = None,
 ) -> None:
     if option.action == "append" or option.nargs == "*":
         append_target = namespace.setdefault(option.name, [])
@@ -279,6 +284,12 @@ def _assign_value_to_dict(
 
     if option.action in ("store", "store_const"):
         namespace[option.name] = value
+        return
+
+    if issubclass(option.action, GcovrConfigOptionAction):
+        option.action(option.flags, option.name)(
+            None, namespace, value, config=cfg_entry_key
+        )
         return
 
     assert False, f"unexpected action for {option.name}: {option.action!r}"
@@ -309,6 +320,18 @@ def merge_options_and_set_defaults(
         target.setdefault(option.name, option.default)
 
     return Options(**target)
+
+
+class UseSortUncoveredNumberAction(GcovrDeprecatedConfigOptionAction):
+    option = "--sort-key"
+    config = "sort-key"
+    value = "uncovered-number"
+
+
+class UseSortUncoveredPercentAction(GcovrDeprecatedConfigOptionAction):
+    option = "--sort-key"
+    config = "sort-key"
+    value = "uncovered-percent"
 
 
 GCOVR_CONFIG_OPTION_GROUPS = [
@@ -490,36 +513,56 @@ GCOVR_CONFIG_OPTIONS = [
         action="store_true",
     ),
     GcovrConfigOption(
-        "sort_uncovered",
+        "sort_key",
+        ["--sort"],
+        config="sort",
+        group="output_options",
+        help=(
+            "Sort entries by filename, number or percent of uncovered lines or branches"
+            "(if the option --sort-branches is given). "
+            "The default order is increasing and can be changed by --sort-reverse. "
+            "The secondary sort key (if values are identical) is always the ascending filename. "
+            "For CSV, HTML, JSON, LCOV and text report."
+        ),
+        choices=["filename", "uncovered-number", "uncovered-percent"],
+        default="filename",
+    ),
+    GcovrConfigOption(
+        "sort_key",
         ["-u", "--sort-uncovered"],
         group="output_options",
         help=(
+            "Deprecated, please use '--sort-key uncovered-number' instead. "
             "Sort entries by number of uncovered lines or branches (if the option "
             "--sort-branches is given). "
             "The default order is increasing and can be changed by --sort-reverse. "
             "The secondary sort key (if values are identical) is always the ascending filename. "
-            "For CSV, HTML, JSON and text report."
+            "For CSV, HTML, JSON, LCOV and text report."
         ),
-        action="store_true",
+        nargs=0,
+        action=UseSortUncoveredNumberAction,
     ),
     GcovrConfigOption(
-        "sort_percent",
+        "sort_key",
         ["-p", "--sort-percentage"],
         group="output_options",
         help=(
+            "Deprecated, please use '--sort-key uncovered-percent' instead. "
             "Sort entries by percentage of uncovered lines or branches (if the option "
             "--sort-branches is given). "
-            "The default order is increasing and can be changed by --sort-reverse."
+            "The default order is increasing and can be changed by --sort-reverse. "
             "The secondary sort key (if values are identical) is always the ascending filename. "
-            "For CSV, HTML, JSON and text report."
+            "For CSV, HTML, JSON, LCOV and text report."
         ),
-        action="store_true",
+        nargs=0,
+        action=UseSortUncoveredPercentAction,
     ),
     GcovrConfigOption(
         "sort_reverse",
         ["--sort-reverse"],
+        config="sort_reverse",
         group="output_options",
-        help="Sort entries in reverse order. For CSV, HTML, JSON and text report.",
+        help="Sort entries in reverse order (see --sort).",
         action="store_true",
     ),
     *formats.get_options(),