+++ gcovr/formats/html/write.py
@@ -44,7 +44,6 @@ from ...utils import (
     force_unix_separator,
     get_md5_hexdigest,
     open_text_for_writing,
-    realpath,
 )
 from ...version import __version__
 
@@ -322,7 +321,9 @@ class RootInfo:
         }
 
         display_filename = force_unix_separator(
-            os.path.relpath(realpath(cdata_fname), realpath(self.directory))
+            os.path.relpath(
+                os.path.realpath(cdata_fname), os.path.realpath(self.directory)
+            )
         )
 
         if link_report is not None:
@@ -432,10 +433,9 @@ def write_report(covdata: CovData, output_file: str, options: Options) -> None:
     filtered_fname = ""
     sorted_keys = sort_coverage(
         covdata,
-        by_branch=options.sort_branches,
-        by_num_uncovered=options.sort_uncovered,
-        by_percent_uncovered=options.sort_percent,
-        reverse=options.sort_reverse,
+        sort_key=options.sort_key,
+        sort_reverse=options.sort_reverse,
+        by_metric="branch" if options.sort_branches else "line",
         filename_uses_relative_pathname=True,
     )
 
@@ -674,10 +674,9 @@ def write_directory_pages(
 
         sorted_files = sort_coverage(
             directory.children,
-            by_branch=options.sort_branches,
-            by_num_uncovered=options.sort_uncovered,
-            by_percent_uncovered=options.sort_percent,
-            reverse=options.sort_reverse,
+            sort_key=options.sort_key,
+            sort_reverse=options.sort_reverse,
+            by_metric="branch" if options.sort_branches else "line",
             filename_uses_relative_pathname=True,
         )