+++ gcovr/options.py
@@ -19,10 +19,10 @@
 
 from __future__ import annotations
 from argparse import ArgumentTypeError
+import argparse
 import logging
 from typing import Any, List, Optional, Union, Callable
 import os
-
 
 LOGGER = logging.getLogger("gcovr")
 
@@ -188,6 +188,31 @@ class Options(object):
         return self.__dict__.get(name)
 
 
+class GcovrConfigOptionAction(argparse.Action):
+    pass
+
+
+class GcovrDeprecatedConfigOptionAction(GcovrConfigOptionAction):
+    def __init__(self, option_strings, dest, **kwargs):
+        super().__init__(option_strings, dest, **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None, config=None):
+        if option_string is not None:
+            LOGGER.warning(
+                f"Deprecated option {option_string} used, please use '{self.option} {self.value}' instead."
+            )
+        if config is not None:
+            LOGGER.warning(
+                f"Deprecated config key {config} used, please use '{self.config}={self.value}' instead."
+            )
+        # This part is used when merging configurations
+        if isinstance(namespace, dict):
+            namespace[self.dest] = values
+        # We are called from argparse
+        else:
+            setattr(namespace, self.dest, self.value)
+
+
 class GcovrConfigOption:
     # pylint: disable=too-many-instance-attributes
     # pylint: disable=too-few-public-methods
@@ -317,7 +342,9 @@ class GcovrConfigOption:
             const = False
             default = True
 
-        assert action in ("store", "store_const", "append")
+        assert action in ("store", "store_const", "append") or issubclass(
+            action, GcovrConfigOptionAction
+        )
 
         self.name = name
         self.flags = flags