+++ gcovr/tests/coexisting_object_directories-from_build_dir/source/fn_04.cpp
@@ -1,0 +1,4 @@
+#include <iostream>
+void fn_04() {
+   std::cout << "04" << std::endl;
+}