+++ gcovr/tests/coexisting_object_directories-from_root_dir-without_object_dir/source/fn_01.cpp
@@ -1,0 +1,4 @@
+#include <iostream>
+void fn_01() {
+   std::cout << "01" << std::endl;
+}