+++ gcovr/tests/coexisting_object_directories-from_root_dir/source/fn_05.cpp
@@ -1,0 +1,4 @@
+#include <iostream>
+void fn_05() {
+   std::cout << "05" << std::endl;
+}