+++ noxfile.py
@@ -23,9 +23,7 @@ import os
 import platform
 import re
 import shutil
-import shlex
 import subprocess
-import sys
 import zipfile
 import nox
 
@@ -500,13 +498,6 @@ def docker_run_compiler(session: nox.Session, version: str) -> None:
     """Run the docker container for a specific GCC version."""
     set_environment(session, version, False)
 
-    def shell_join(args):
-        if sys.version_info >= (3, 8):
-            return shlex.join(args)
-        else:
-            # Code for join taken from Python 3.9
-            return " ".join(shlex.quote(arg) for arg in args)
-
     nox_options = session.posargs
     if not session.interactive:
         nox_options.insert(0, "--non-interactive")
@@ -633,12 +624,16 @@ def import_reference(session: nox.Session) -> None:
             "Exact one ZIP file needed. Usage: nox -s import_reference -- file.zip"
         )
 
+    def extract(fh_zip: zipfile.ZipFile):
+        for entry in fh_zip.filelist:
+            session.log(fh_zip.extract(entry, "gcovr/tests"))
+
     with zipfile.ZipFile(session.posargs[0]) as fh_zip:
         try:
             zip_info_diff_zip = fh_zip.getinfo("diff.zip")
             with fh_zip.open(zip_info_diff_zip) as fh_inner_zip:
                 seekable_buf = io.BytesIO(fh_inner_zip.read())
                 with zipfile.ZipFile(seekable_buf) as fh_diff_zip:
-                    fh_diff_zip.extractall("gcovr/tests")
+                    extract(fh_diff_zip)
         except KeyError:
-            fh_zip.extractall("gcovr/tests")
+            extract(fh_zip)