+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Known bugs:
 
 Breaking changes:
 
+- Dropped support for Python 3.7 (:issue:`869`)
 - The exit code for an error of the reader module is changed from 8 to 64 and for a writer from 7 to 128. (:issue:`773`)
 
 New features and notable changes:
@@ -27,10 +28,12 @@ New features and notable changes:
 - Add :option:`--jacoco` to generate JaCoCo XML format. (:issue:`823`))
 - Add function coverage to HTML report. (:issue:`828`)
 - Improve sorting of data in reports (:issue:`817`):
+
   - Sort file names alpha numerical and with casefold (see `str.casefold <https://docs.python.org/3.11/library/stdtypes.html?highlight=str%20casefold#str.casefold>`_) (``file_10.c`` comes after ``file_0.c``).
   - Always sort at the end by filename if line or branch coverage is identical for a file.
   - Add :option:`--sort-branches` to sort by branches instead of lines, this is the default if :option:`--txt-branches` is used.
   - Add :option:`--sort-reverse` to reverse the sort order.
+
 - Add option to report covered lines in txt report. (:issue:`836`)
 - Add support for specifying files for :option:`search_paths`. (:issue:`834`)
 - Use different color for partial covered lines in HTML report. (:issue:`839`)
@@ -38,6 +41,14 @@ New features and notable changes:
 - Add support for FIPS enabled OS when used with Python 3.9. (:issue:`850`)
 - Ignore all negative hits if :option:`--gcov-ignore-parse-errors` is used. (:issue:`852`)
 - Reduce file size for detailed HTML reports by merging columns the function lists. (:issue:`840`)
+- Use literal options for sorting and TXT metric. (:issue:`867`)
+
+  - The :option:`-b`, :option:`--txt-branches` and :option:`--branches` are deprecated, use :option:`--txt-metric` instead.
+    The reason for this is that we have line, branch and decision coverage and handle this with flags is more complex than
+    using an enumeration.
+  - The :option:`--sort-uncovered` and :option:`--sort-percentage` are deprecated, use :option:`--sort` instead.
+    The reason for this is that only one sorting order shall be selectable and and an enumeration is easier to handle
+    than several flags.
 - Split list of functions into tables with maximum 10000 rows to fix rendering issues. (:issue:`858`)
 
 Bug fixes and small improvements:
@@ -82,6 +93,7 @@ Internal changes:
 - Use ``build`` instead of calling ``setup.py`` directly. (:issue:`819`)
 - Add nox session to import reference file from pipeline. (:issue:`831`)
 - Add support for ``clang-15`` in our test suite and fix test with write protection under Mac OS. (:issue:`853`)
+- Add test for parallel execution of multiple gcovr instances. (:issue:`832`)
 
 6.0 (08 March 2023)
 -------------------